Update for the voting part, css

+++ database/migrations/2016_04_13_080619_add_columns_to_answer.php
@@ -13,7 +13,7 @@ class AddColumnsToAnswer extends Migration
     public function up()
     {
         Schema::table('answer', function (Blueprint $table) {
-            $table->integer('note_average');
+            $table->float('note_average');
         });
     }
